feat(authz): Implement role-based authorization using Policies - restrict delete actions to admins

+++ app/Http/Controllers/Api/CampaignController.php
@@ -74,6 +74,7 @@ class CampaignController extends Controller
      */
     public function destroy(Campaign $campaign)
     {
+        $this->authorize('delete', $newsletter);
         $campaign->delete();
 
         return response()->json(null, 204);
